refactor: 구인 공고 리스트 페이지 코드 구조 개선
