+++ src/components/commonInProject/SubHeader/SubHeader.tsx
@@ -18,7 +18,7 @@ const SubHeader = ({
   return (
     <Hstack
       gap="lg"
-      className="pb-oz-xxl h-[100px] w-full max-w-[1216px] items-center justify-between"
+      className="pb-oz-xxl px-oz-xxl h-[100px] w-full items-center justify-between"
     >
       {isBackButtonVisible && (
         <button
